+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixes
 
+- `[expect]` Display expectedDiff more carefully in toBeCloseTo ([#8389](https://github.com/facebook/jest/pull/8389))
 - `[jest-fake-timers]` `getTimerCount` will not include cancelled immediates ([#8764](https://github.com/facebook/jest/pull/8764))
 
 ### Chore & Maintenance